+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Get-SsisOperation command.
+- Get-SsisExecutionMessage command.
 - Wait-SsisExecution command.
 - Stop-SsisExecution command.
 - Start-SsisExecution command.